[Refacotr] Move ArrayExpr from analysis to sql/ast
