+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/ArrayTypeTest.java
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-
 package com.starrocks.sql.plan;
 
 import com.google.common.collect.Sets;
